Obtain and format results produced by tuning

Obtain and format results produced by tuning

collect_predictions(x, nest_by = "nothing", wflow = NULL)

collect_metrics(x, nest_by = "nothing", summarize = TRUE, wflow = NULL)

Arguments

x
The results of tune_grid() or tune_bayes(). For collect_predictions(), the control option save_pred = TRUE should have been used.

nest_by
A character string to indicate which (if any) columns should be in the resulting tibble and which should be nested. Values are "nothing" (i.e. no nesting), "resamples" (nest the parameter values), and "parameters" (nest the resamples).

wflow
The workflow associated with the results (only required for nest_by = "parameters").

summarize
A logical; should metrics be summarized over resamples (TRUE) or return the values for each individual resample. Only appropriate when nest_by = "nothing".

Value

A tibble. The column names depend on the results and the mode of the model.

Examples

data("example_ames_knn")

# Summarized over resamples
collect_metrics(ames_grid_search)

#> # A tibble: 20 x 10
#>        K weight_func dist_power   lon   lat .metric .estimator   mean     n
#>    <int> <chr>            <dbl> <int> <int> <chr>   <chr>       <dbl> <int>
#>  1     3 gaussian         1.86     10    15 rmse    standard   0.0848    10
#>  2     3 gaussian         1.86     10    15 rsq     standard   0.777     10
#>  3     4 biweight         0.311     9     6 rmse    standard   0.0777    10
#>  4     4 biweight         0.311     9     6 rsq     standard   0.813     10
#>  5     5 gaussian         0.411     3     8 rmse    standard   0.0744    10
#>  6     5 gaussian         0.411     3     8 rsq     standard   0.833     10
#>  7    12 epanechnik…      1.53      6     8 rmse    standard   0.0785    10
#>  8    12 epanechnik…      1.53      6     8 rsq     standard   0.815     10
#>  9    21 triweight        0.626     3     6 rmse    standard   0.0736    10
#> 10    21 triweight        0.626     3     6 rsq     standard   0.836     10
#> 11    32 triangular       0.165    10    15 rmse    standard   0.0771    10
#> 12    32 triangular       0.165    10    15 rsq     standard   0.826     10
#> 13    33 triweight        0.511    11     5 rmse    standard   0.0733    10
#> 14    33 triweight        0.511    11     5 rsq     standard   0.840     10
#> 15    35 gaussian         1.29      5    14 rmse    standard   0.0818    10
#> 16    35 gaussian         1.29      5    14 rsq     standard   0.815     10
#> 17    35 rank             1.32      9     3 rmse    standard   0.0821    10
#> 18    35 rank             1.32      9     3 rsq     standard   0.813     10
#> 19    40 triangular       0.167    12     8 rmse    standard   0.0782    10
#> 20    40 triangular       0.167    12     8 rsq     standard   0.822     10
#> # … with 1 more variable: std_err <dbl>

# Per-resample values
collect_metrics(ames_grid_search, summarize = FALSE)

#> # A tibble: 200 x 9
#>    id         K weight_func dist_power   lon   lat .metric .estimator .estimate
#>    <chr>  <int> <chr>            <dbl> <int> <int> <chr>   <chr>          <dbl>
#>  1 Fold01    35 rank             1.32      9     3 rmse    standard      0.0885
#>  2 Fold01    35 rank             1.32      9     3 rsq     standard      0.808 
#>  3 Fold01    35 gaussian         1.29      5    14 rmse    standard      0.0887
#>  4 Fold01    35 gaussian         1.29      5    14 rsq     standard      0.806 
#>  5 Fold01    21 triweight        0.626     3     6 rmse    standard      0.0811
#>  6 Fold01    21 triweight        0.626     3     6 rsq     standard      0.817 
#>  7 Fold01     4 biweight         0.311     9     6 rmse    standard      0.0892
#>  8 Fold01     4 biweight         0.311     9     6 rsq     standard      0.775 
#>  9 Fold01     3 gaussian         1.86     10    15 rmse    standard      0.0891
#> 10 Fold01     3 gaussian         1.86     10    15 rsq     standard      0.776 
#> # … with 190 more rows