Control the Bayesian search process

Control the Bayesian search process

ctrl_bayes(verbose = FALSE, no_improve = 10L, uncertain = Inf,
  seed = sample.int(10^5, 1), extract = NULL, save_pred = FALSE,
  time_limit = NA, pkgs = NULL)

Arguments

verbose
A logical for logging results as they are generated. Despite this argument, warnings and errors are always shown.

no_improve
The integer cutoff for the number of iterations without better results.

uncertain
The number of iterations with no improvement before an uncertainty sample is created where a sample with high predicted variance is chosen. The iteration counter is reset after each uncertainty sample. For example, if uncertain = 10, this condition is triggered every 10 samples with no improvement.

seed
An integer for controlling the random number stream.

extract
An optional function to collect any information from the model or other objects. See ctrl_grid() for details. Note that if initial results were already generated using tune_grid(), care must be taken if the Bayesian search has a different extraction function.

save_pred
A logical to save the out-of-sample predictions from each resample and each parameter combination. See ctrl_grid() for details.

time_limit
A number for the minimum number of minutes (elapsed) that the function should execute. The elapsed time is evaluated at internal checkpoints and, if over time, the results at that time are returned (with a warning). This means that the time_limit is not an exact limit, but a minimum time limit.

pkgs
An optional character string of R package names that should be loaded (by namespace) during parallel processing.