Control the grid search process

Control the grid search process

grid_control(verbose = FALSE, allow_par = TRUE, extract = NULL,
  save_pred = FALSE, pkgs = NULL)

Arguments

verbose
A logical for logging results as they are generated. Despite this argument, warnings and errors are always shown.

allow_par
A logical to allow parallel processing (if a parallel backend is registered).

extract
An optional function with at least one argument (or NULL) that can be used to retain arbitrary objects from the model fit object, recipe, or other elements of the workflow.

save_pred
A logical for whether the out-of-sample predictions should be saved for each model evaluated.

pkgs
An optional character string of R package names that should be loaded (by namespace) during parallel processing.

Details

For extract, this function can be used to output the model object, the recipe (if used), or some components of either or both. When evaluated, the function's sole argument has a named list with elements recipe and model. If the formula method is used, the recipe element will be NULL.

The results of the extract function are added to a list column in the output called .extracts. Each element of this list is a tibble with tuning parameter column and a list column (also called .extracts) that contains the results of the function. If no extraction function is used, there is no .extracts column in the resulting object.