Fine Foods example data

Fine Foods example data

Source

https://snap.stanford.edu/data/web-FineFoods.html

Value

training_data,testing_data

tibbles

Details

These data are from Amazon, who describe it as "This dataset consists of reviews of fine foods from amazon. The data span a period of more than 10 years, including all ~500,000 reviews up to October 2012. Reviews include product and user information, ratings, and a plaintext review."

A subset of the data are contained here and are split into a training and test set. The training set sampled 10 products and retained all of their individual reviews. Since the reviews within these products are correlated, we recommend resampling the data using a leave-one-product-out approach. The test set sampled 500 products that were not included in the training set and selected a single review at random for each.

There is a column for the product, a column for the text of the review, and a factor column for a class variable. The outcome is whether the reviewer gave the product a 5-star rating or not.

Examples

data(small_fine_foods)
str(training_data)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	2137 obs. of  3 variables:
#>  $ product: chr  "B005HGAV8I" "B005HGAV8I" "B005HGAV8I" "B005HGAV8I" ...
#>  $ review : chr  "I ordered the 35 pack and received two of each flavor, and three of one flavor.  I understand this is a random "| __truncated__ "I ordered the 35 pack.  Maybe I just got lucky but I had an excellent variety!  I have 19 different kinds, doub"| __truncated__ "I received the Crazy Cups Flavored Coffee Sampler (pack of 35) today and am quite pleased with it.  The K-cups "| __truncated__ "The price isn't too bad, but I guess my expectations were too high.  Most of these coffees were very low on fla"| __truncated__ ...
#>  $ score  : Factor w/ 2 levels "great","other": 2 1 1 2 2 1 2 1 2 1 ...