A placeholder function for argument values that are to be tuned.

tune() is used when a parameter will be specified at a later date.

tune(id = "")

Arguments

id
A single character value that can be used to differentiate parameters that are used in multiple places but have the same name, or if the user wants a note associated with the parameter.

Value

A call object that echos the user input.

See also

Examples

tune()

#> tune()

class(tune())

#> [1] "call"

tune("your name here")

#> tune("your name here")

# How `tune()` is used in practice:

library(parsnip)
nearest_neighbor(
  mode = "regression",
  neighbors = tune("K"),
  weight_func = tune(),
  dist_power = tune()
)

#> K-Nearest Neighbor Model Specification (regression)
#> 
#> Main Arguments:
#>   neighbors = tune("K")
#>   weight_func = tune()
#>   dist_power = tune()
#>