Obtain and format results produced by tuning functions

Obtain and format results produced by tuning functions

collect_predictions(x)

collect_metrics(x, summarize = TRUE)

Arguments

x
The results of tune_grid(), tune_bayes(), fit_resamples(), or last_fit(). For collect_predictions(), the control option save_pred = TRUE should have been used.

summarize
A logical; should metrics be summarized over resamples (TRUE) or return the values for each individual resample. Note that, if x is created by last_fit(), summarize has not effect.

Value

A tibble. The column names depend on the results and the mode of the model.

For collect_metrics() and collect_predictions(), there are columns for each tuning parameter (using the id from tune(), if any).

For collect_predictions(), there are additional columns for the resampling identifier(s), columns for the predicted values (e.g., .pred, .pred_class, etc.), and a column for the outcome(s) using the original column name(s) in the data.

collect_metrics() also has columns .metric, and .estimator. When the results are summarized, there are columns for mean, n, and std_err. When not summarized, the additional columns for the resampling identifier(s) and .estimate.

Examples

data("example_ames_knn")

# Summarized over resamples
collect_metrics(ames_grid_search)

#> # A tibble: 20 x 10
#>        K weight_func dist_power   lon   lat .metric .estimator   mean     n
#>    <int> <chr>            <dbl> <int> <int> <chr>   <chr>       <dbl> <int>
#>  1     3 gaussian         1.86     10    15 rmse    standard   0.0867    10
#>  2     3 gaussian         1.86     10    15 rsq     standard   0.766     10
#>  3     4 biweight         0.311     9     6 rmse    standard   0.0782    10
#>  4     4 biweight         0.311     9     6 rsq     standard   0.811     10
#>  5     5 gaussian         0.411     3     8 rmse    standard   0.0740    10
#>  6     5 gaussian         0.411     3     8 rsq     standard   0.832     10
#>  7    12 epanechnik…      1.53      6     8 rmse    standard   0.0772    10
#>  8    12 epanechnik…      1.53      6     8 rsq     standard   0.821     10
#>  9    21 triweight        0.626     3     6 rmse    standard   0.0732    10
#> 10    21 triweight        0.626     3     6 rsq     standard   0.837     10
#> 11    32 triangular       0.165    10    15 rmse    standard   0.0769    10
#> 12    32 triangular       0.165    10    15 rsq     standard   0.826     10
#> 13    33 triweight        0.511    11     5 rmse    standard   0.0726    10
#> 14    33 triweight        0.511    11     5 rsq     standard   0.842     10
#> 15    35 gaussian         1.29      5    14 rmse    standard   0.0809    10
#> 16    35 gaussian         1.29      5    14 rsq     standard   0.816     10
#> 17    35 rank             1.32      9     3 rmse    standard   0.0811    10
#> 18    35 rank             1.32      9     3 rsq     standard   0.813     10
#> 19    40 triangular       0.167    12     8 rmse    standard   0.0779    10
#> 20    40 triangular       0.167    12     8 rsq     standard   0.822     10
#> # … with 1 more variable: std_err <dbl>

# Per-resample values
collect_metrics(ames_grid_search, summarize = FALSE)

#> # A tibble: 200 x 9
#>    id         K weight_func dist_power   lon   lat .metric .estimator .estimate
#>    <chr>  <int> <chr>            <dbl> <int> <int> <chr>   <chr>          <dbl>
#>  1 Fold01    35 rank             1.32      9     3 rmse    standard      0.0879
#>  2 Fold01    35 rank             1.32      9     3 rsq     standard      0.804 
#>  3 Fold01    35 gaussian         1.29      5    14 rmse    standard      0.0873
#>  4 Fold01    35 gaussian         1.29      5    14 rsq     standard      0.808 
#>  5 Fold01    21 triweight        0.626     3     6 rmse    standard      0.0825
#>  6 Fold01    21 triweight        0.626     3     6 rsq     standard      0.810 
#>  7 Fold01     4 biweight         0.311     9     6 rmse    standard      0.0885
#>  8 Fold01     4 biweight         0.311     9     6 rsq     standard      0.779 
#>  9 Fold01     3 gaussian         1.86     10    15 rmse    standard      0.0951
#> 10 Fold01     3 gaussian         1.86     10    15 rsq     standard      0.743 
#> # … with 190 more rows

# ---------------------------------------------------------------------------

library(parsnip)
library(rsample)

#> Loading required package: tidyr

lm_mod <-linear_reg() %>% set_engine("lm")
set.seed(93599150)
car_folds <- vfold_cv(mtcars)
ctrl <- control_resamples(save_pred = TRUE)

resampled <- fit_resamples(mpg ~ ., lm_mod, resamples = car_folds, control = ctrl)
collect_predictions(resampled)

#> # A tibble: 32 x 4
#>    id     .pred  .row   mpg
#>    <chr>  <dbl> <int> <dbl>
#>  1 Fold01  30.8    19  30.4
#>  2 Fold01  25.6    21  21.5
#>  3 Fold01  17.4    30  19.7
#>  4 Fold01  10.1    31  15  
#>  5 Fold02  22.5     2  21  
#>  6 Fold02  13.5    12  16.4
#>  7 Fold02  16.2    25  19.2
#>  8 Fold02  28.5    26  27.3
#>  9 Fold03  20.9    11  17.8
#> 10 Fold03  27.0    28  30.4
#> # … with 22 more rows