Extract elements of tune objects

These functions extract various elements from a tune object. If they do not exist yet, an error is thrown.

extract_preprocessor() returns the formula, recipe, or variable expressions used for preprocessing.

extract_spec_parsnip() returns the parsnip model specification.

extract_fit_parsnip() returns the parsnip model fit object.

extract_fit_engine() returns the engine specific fit embedded within a parsnip model fit. For example, when using parsnip::linear_reg() with the "lm" engine, this returns the underlying lm object.

extract_mold() returns the preprocessed "mold" object returned from hardhat::mold(). It contains information about the preprocessing, including either the prepped recipe, the formula terms object, or variable selectors.

extract_recipe() returns the recipe. The estimated argument specifies whether the fitted or original recipe is returned.

extract_workflow() returns the workflow object if the control option save_workflow = TRUE was used. The workflow will only have been estimated for objects produced by last_fit().

# S3 method for last_fit
extract_workflow(x, ...)

# S3 method for tune_results
extract_workflow(x, ...)

# S3 method for tune_results
extract_spec_parsnip(x, ...)

# S3 method for tune_results
extract_recipe(x, ..., estimated = TRUE)

# S3 method for tune_results
extract_fit_parsnip(x, ...)

# S3 method for tune_results
extract_fit_engine(x, ...)

# S3 method for tune_results
extract_mold(x, ...)

# S3 method for tune_results
extract_preprocessor(x, ...)

Arguments

x
A tune_results object.

...
Not currently used.

estimated
A logical for whether the original (unfit) recipe or the fitted recipe should be returned.

Value

The extracted value from the tune tune_results, x, as described in the description section.

Details

These functions supersede extract_model().

Examples

library(recipes)
library(rsample)
library(parsnip)

set.seed(6735)
tr_te_split <- initial_split(mtcars)

spline_rec <- recipe(mpg ~ ., data = mtcars) %>%
  step_ns(disp)

lin_mod <- linear_reg() %>%
  set_engine("lm")

spline_res <- last_fit(lin_mod, spline_rec, split = tr_te_split)

extract_preprocessor(spline_res)

#> Data Recipe
#> 
#> Inputs:
#> 
#>       role #variables
#>    outcome          1
#>  predictor         10
#> 
#> Operations:
#> 
#> Natural Splines on disp

# The `spec` is the parsnip spec before it has been fit.
# The `fit` is the fitted parsnip model.
extract_spec_parsnip(spline_res)

#> Linear Regression Model Specification (regression)
#> 
#> Computational engine: lm 
#> 

extract_fit_parsnip(spline_res)

#> parsnip model object
#> 
#> Fit time:  2ms 
#> 
#> Call:
#> stats::lm(formula = ..y ~ ., data = data)
#> 
#> Coefficients:
#> (Intercept)          cyl           hp         drat           wt         qsec  
#>    3.543425     2.061709    -0.008018     0.932096    -4.700860     0.935636  
#>          vs           am         gear         carb    disp_ns_1    disp_ns_2  
#>    1.053248    -0.942407     1.932864    -0.477392   -15.305563    -1.806417  
#> 

extract_fit_engine(spline_res)

#> 
#> Call:
#> stats::lm(formula = ..y ~ ., data = data)
#> 
#> Coefficients:
#> (Intercept)          cyl           hp         drat           wt         qsec  
#>    3.543425     2.061709    -0.008018     0.932096    -4.700860     0.935636  
#>          vs           am         gear         carb    disp_ns_1    disp_ns_2  
#>    1.053248    -0.942407     1.932864    -0.477392   -15.305563    -1.806417  
#> 

# The mold is returned from `hardhat::mold()`, and contains the
# predictors, outcomes, and information about the preprocessing
# for use on new data at `predict()` time.
extract_mold(spline_res)

#> $predictors
#> # A tibble: 24 x 11
#>      cyl    hp  drat    wt  qsec    vs    am  gear  carb disp_ns_1 disp_ns_2
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>     <dbl>     <dbl>
#>  1     8   245  3.73  3.84  15.4     0     0     3     4    0.528     0.170 
#>  2     8   180  3.07  3.73  17.6     0     0     3     3    0.575    -0.110 
#>  3     6   110  3.9   2.88  17.0     0     1     4     4    0.370    -0.199 
#>  4     8   150  3.15  3.44  17.3     0     0     3     2    0.570    -0.0191
#>  5     8   335  3.54  3.57  14.6     0     1     5     8    0.571    -0.0298
#>  6     4    95  3.92  3.15  22.9     1     0     4     2    0.300    -0.168 
#>  7     8   245  3.21  3.57  15.8     0     0     3     4    0.514     0.217 
#>  8     4    66  4.08  2.2   19.5     1     1     4     1    0.0343   -0.0203
#>  9     4    62  3.69  3.19  20       1     0     4     2    0.322    -0.178 
#> 10     8   180  3.07  4.07  17.4     0     0     3     3    0.575    -0.110 
#> # … with 14 more rows
#> 
#> $outcomes
#> # A tibble: 24 x 1
#>      mpg
#>    <dbl>
#>  1  13.3
#>  2  17.3
#>  3  21  
#>  4  15.2
#>  5  15  
#>  6  22.8
#>  7  14.3
#>  8  32.4
#>  9  24.4
#> 10  16.4
#> # … with 14 more rows
#> 
#> $blueprint
#> Recipe blueprint: 
#>  
#> # Predictors: 10 
#>   # Outcomes: 1 
#>    Intercept: FALSE 
#> Novel Levels: FALSE 
#>  Composition: tibble 
#> 
#> $extras
#> $extras$roles
#> NULL
#> 
#> 

# A useful shortcut is to extract the fitted recipe from the workflow
extract_recipe(spline_res)

#> Data Recipe
#> 
#> Inputs:
#> 
#>       role #variables
#>    outcome          1
#>  predictor         10
#> 
#> Training data contained 24 data points and no missing data.
#> 
#> Operations:
#> 
#> Natural Splines on disp [trained]

# That is identical to
identical(
  extract_mold(spline_res)$blueprint$recipe,
  extract_recipe(spline_res)
)

#> [1] TRUE