Determine the minimum set of model fits

min_grid determines exactly what models should be fit in order to evaluate the entire set of tuning parameter combinations. This is for internal use only and the API may change in the near future.

min_grid(x, grid, ...)

# S3 method for model_spec
min_grid(x, grid, ...)

# S3 method for boost_tree
min_grid(x, grid, ...)

# S3 method for linear_reg
min_grid(x, grid, ...)

# S3 method for logistic_reg
min_grid(x, grid, ...)

# S3 method for mars
min_grid(x, grid, ...)

# S3 method for multinom_reg
min_grid(x, grid, ...)

# S3 method for nearest_neighbor
min_grid(x, grid, ...)

Arguments

x
A model specification.

grid
A tibble with tuning parameter combinations.

...
Not currently used.

Value

A tibble with the minimum tuning parameters to fit and an additional list column with the parameter combinations used for prediction.