Obtain the numerically best parameters

select_best() finds the tuning parameter combination with the best performance values. show_best() displays the top sub-models and their performance estimates.

show_best(x, metric, n_top = 5, maximize = TRUE)

select_best(x, metric, maximize = TRUE)

Arguments

x
The results of tune_grid() or tune_Bayes().

metric
A character value for the metric that will be used to sort the models. (See https://tidymodels.github.io/yardstick/articles/metric-types.html for more details). Not required if a single metric exists in x.

n_top
An integer for the number of top results/rows to return.

maximize
A logical value (TRUE/FALSE).

Value

A tibble with columns for the parameters. show_best() also includes columns for performance metrics.