Fine Foods example data

Fine Foods example data

Source

https://snap.stanford.edu/data/web-FineFoods.html

Value

training_data,testing_data

tibbles

Details

These data are from Amazon, who describe it as "This dataset consists of reviews of fine foods from amazon. The data span a period of more than 10 years, including all ~500,000 reviews up to October 2012. Reviews include product and user information, ratings, and a plaintext review."

A subset of the data are contained here and are split into a training and test set. The training set sampled 25 products and retained all of their individual reviews. Since the reviews within these products are correlated, we recommend resampling the data using a leave-one-product-out approach. The test set sampled 1,000 products that were not included in the training set and selected a single review at random for each.

There is a column for the product, a column for the text of the review, and a factor column for a class variable. The outcome is whether the reviewer game the product a 5-star rating or not.

Examples

data(small_fine_foods)
str(training_data)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	5334 obs. of  3 variables:
#>  $ product: chr  "B003D4IYSU" "B003D4IYSU" "B003D4IYSU" "B003D4IYSU" ...
#>  $ review : chr  "A well flavored Tea for after dinner or to use when you have trouble sleeping. I have purchaced this enjoyable "| __truncated__ "This is the first time I've tried this brand, this flavor and this type of tea.  It is so good!  With a little "| __truncated__ "This tea is so refreshing, and I just love the taste.  Not a heavy chocolate taste, but really gets the endorph"| __truncated__ "This is delicious hot or cold! There is enough that I could share w/family & friends. The flavor is strong & do"| __truncated__ ...
#>  $ score  : Factor w/ 2 levels "great","other": 1 1 1 1 1 1 1 2 2 1 ...