Bayesian optimization of model parameters.

tune_Bayes() uses models to generate new candidate tuning parameter combinations based on previous results.

tune_Bayes(object, ...)

# S3 method for default
tune_Bayes(object, ...)

# S3 method for recipe
tune_Bayes(object, model, rs, iter = 10,
  param_info = NULL, perf = NULL, objective = exp_improve(),
  initial = NULL, control = Bayes_control(), ...)

# S3 method for formula
tune_Bayes(formula, model, rs, iter = 10,
  param_info = NULL, perf = NULL, objective = exp_improve(),
  initial = NULL, control = Bayes_control(), ...)

# S3 method for workflow
tune_Bayes(object, model = NULL, rs, iter = 10,
  param_info = NULL, perf = NULL, objective = exp_improve(),
  initial = NULL, control = Bayes_control(), ...)

Arguments

object
A model workflow or recipe object.

...
Not currently used.

model
A parsnip model specification (or NULL when object is a workflow).

rs
An rset() object. This argument should be named.

iter
The maximum number of search iterations.

param_info
A dials::param_set() object or NULL. If none is given, a parameters set is derived from other arguments.

perf
A yardstick::metric_set() object containing information on how models will be evaluated for performance. The first metric in perf is the one that will be optimized.

objective
A character string for what metric should be optimized or an acquisition function object.

initial
An initial set of results in a tidy format.

control
A control object created by Bayes_control()

formula
A traditional model formula.

Value

A tibble of results that mirror those generated by tune_grid(). However, these results contain an .iter column and replicate the rset object multiple times over iterations (at limited additional memory costs).

Details

The optimization starts with a set of initial results, such as those generated by tune_grid(). If none exist, the function will create several combinations and obtain their performance estimates.

Using one of the performance estimates as the model outcome, a Gaussian process (GP) model is created where the previous tuning parameter combinations are used as the predictors.

A large grid of potential hyperparameter combinations is predicted using the model and scored using an acquisition function. These functions usually combine the predicted mean and variance of the GP to decide the best parameter combination to try next. For more information, see the documentation for exp_improve() and the corresponding package vignette.

The best combination is evaluated using resampling and the process continues.

TODO notes on when to use, pluses and minuses.